# Optimized CV method (kESI's code)