@abukaj 2023-08-03
The actual improvement may be greater, as `k.update_R()` for n=100 takes 6s of CPU time (3s of Wall time)